requirements.txt
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.9.0
cffi==2.0.0
channels==4.3.2
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
dj-database-url==3.1.2
Django==6.0.3
djangorestframework==3.16.1
drf-spectacular==0.29.0
greenlet==3.3.2
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==26.0
paho-mqtt==2.1.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyOpenSSL==26.0.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
service-identity==24.2.0
SQLAlchemy==2.0.48
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
uritemplate==4.2.0
zope.interface==8.2
